fix(PR-1870): fold draft action guard into draft wire validation
