+++ crates/server-core/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod deck_resolve;
-pub mod draft_action_wire_guard;
 pub mod draft_session;
 pub mod draft_wire_guard;
 pub mod emote_guard;
@@ -18,10 +17,10 @@ pub mod spectator_wire_guard;
 pub mod starter_decks;
 
 pub use deck_resolve::resolve_deck;
-pub use draft_action_wire_guard::guard_draft_action;
 pub use draft_session::{generate_draft_code, DraftSession, DraftSessionManager};
 pub use draft_wire_guard::{
-    guard_create_draft_with_settings, guard_join_draft_with_password, guard_reconnect_draft,
+    guard_create_draft_with_settings, guard_draft_action, guard_join_draft_with_password,
+    guard_reconnect_draft,
 };
 pub use emote_guard::guard_emote;
 pub use filter::filter_state_for_player;
